+++ CHANGELOG.md
@@ -9,11 +9,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+### Fixes
+
+### Breaking
+
+## [2.160.0] - 2026-03-18
+
+### Features
+- program: admin withdrawal from IF [#2120](https://github.com/drift-labs/protocol-v2/pull/2120)
+
+### Fixes
+
+### Breaking
+
+## [2.159.0] - 2026-03-16
+### Features
+
 - program: allow for swap-involved ATA closures [#2118](https://github.com/drift-labs/protocol-v2/pull/2118)
+- program: add market config flag to disable formulaic k update [#2136](https://github.com/drift-labs/protocol-v2/pull/2136)
+- program: filter out stink bids/asks from perp twap crank [#2124](https://github.com/drift-labs/protocol-v2/pull/2124)
 
 ### Fixes
 
 - program: prevent user from enabling HLM when they are failing maintenance margin check [#2116](https://github.com/drift-labs/protocol-v2/pull/2116)
+- program: fix bug where users are stuck in liquidation status after completed liquidation [#2122](https://github.com/drift-labs/protocol-v2/pull/2122)
+- program: skip isolated positions when checking for cross margin bankruptcy [#2123](https://github.com/drift-labs/protocol-v2/pull/2123)
 
 ### Breaking